Meta_DPi_Server/requirements.txt
asgiref==3.2.10
biopython==1.77
cycler==0.10.0
Django==3.0.8
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mpld3==0.5.1
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
ProDy==1.10.11
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
